requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==3.2.18
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
environ==1.0
filelock==3.12.0
Markdown==3.4.4
mysqlclient==2.2.0
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
virtualenv==20.22.0
